requirements.txt
biopython==1.80
contourpy==1.0.6
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
html2text==2020.1.16
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pyfastx==0.9.1
pyparsing==3.0.9
pysam==0.20.0
python-dateutil==2.8.2
pytz==2022.7
seaborn==0.12.2
six==1.16.0
snapgene-reader==0.1.20
tqdm==4.64.1
xmltodict==0.13.0
